requirements.txt
Brotli==1.0.4
dj-database-url==0.4.2
Django==2.0.6
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==2.6.11
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
whitenoise==3.3.1
